# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "lister tout les clients ou lister les clients avec compte ou sans compte et lister client avec compte active ou non",
                "operationId": "24bc4620ac92bdf18458e017d5a6307e",
                "parameters": [
                    {
                        "name": "comptes",
                        "in": "query",
                        "description": "Filter clients by comptes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "operationId": "85171dfc3a4ea754d223efa15689306e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "surnom",
                                    "telephone"
                                ],
                                "properties": {
                                    "surnom": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Obtenir les informations d'un client par ID",
                "description": "Récupère les informations d'un client en utilisant son ID.",
                "operationId": "e71e872edf8bbb1364990cc3d2c2d99c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client trouvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID du client",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Nom du client",
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "description": "Email du client",
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "phone": {
                                                    "description": "Numéro de téléphone du client",
                                                    "type": "string",
                                                    "example": "123-456-7890"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "Message de réponse",
                                            "type": "string",
                                            "example": "Client trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "description": "Message de réponse",
                                            "type": "string",
                                            "example": "Objet non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clients/{id}/dettes": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Obtenir les dettes d'un client par ID",
                "description": "Récupère les informations d'un client ainsi que ses dettes en utilisant son ID.",
                "operationId": "d436fa683d5e4f44c8adbccf9866d823",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client et dettes trouvés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID du client",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Nom du client",
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "description": "Email du client",
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "phone": {
                                                    "description": "Numéro de téléphone du client",
                                                    "type": "string",
                                                    "example": "123-456-7890"
                                                },
                                                "dettes": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID de la dette",
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "amount": {
                                                                "description": "Montant de la dette",
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 100
                                                            },
                                                            "due_date": {
                                                                "description": "Date d'échéance de la dette",
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-12-31"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "Message de réponse",
                                            "type": "string",
                                            "example": "Client et dettes trouvés"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "description": "Message de réponse",
                                            "type": "string",
                                            "example": "Objet non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clients/telephone": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Rechercher un client par numéro de téléphone",
                "operationId": "7a96a40fa7f17ab4cff2d7da8f570dd4",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "Numéro de téléphone du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clients trouvés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clients": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clients trouvés"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun client trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clients": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun client trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "surnom": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "photo": {
                        "type": "string",
                        "example": "path/to/photo.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authorization header using the Bearer scheme."
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}